Working confluence uploader
Finished a working implementation of the confluence uploader.
Refactored oneNote.py as a generator to generate docBlocks as it traverses the file tree.
Added more try except blocks to catch errors during runtime.
Modified session block for confluence allowing for quicker uploads to confluence using a backoff factor of 0.1 to deal with rate limiting.
Set max retries for confluence connection session to 1000.

If one note exporter does not complete a full upload to mongo, it should drop all partial progress. This is